--- a/pkg/parquetquery/predicates.go
+++ b/pkg/parquetquery/predicates.go
@@ -133,20 +133,23 @@ func (p *regexPredicate) keep(v *pq.Value) bool {
 		return false
 	}
 
-	s := v.String()
-	if matched, ok := p.matches[s]; ok {
+	// Check uses zero alloc optimization of map[string([]byte)]
+	b := v.ByteArray()
+	if matched, ok := p.matches[string(b)]; ok {
 		return matched
 	}
 
 	matched := false
 	for _, r := range p.regs {
-		if r.MatchString(s) == p.shouldMatch {
+		if r.Match(b) == p.shouldMatch {
 			matched = true
 			break
 		}
 	}
 
-	p.matches[s] = matched
+	// Only alloc the string when updating the map
+	p.matches[v.String()] = matched
+
 	return matched
 }