--- a/pkg/parquetquery/predicates.go
+++ b/pkg/parquetquery/predicates.go
@@ -133,8 +133,9 @@ func (p *regexPredicate) keep(v *pq.Value) bool {
 		return false
 	}
 
-	// Check uses zero alloc optimization of map[string([]byte)]
 	b := v.ByteArray()
+
+	// Check uses zero alloc optimization of map[string([]byte)]
 	if matched, ok := p.matches[string(b)]; ok {
 		return matched
 	}
@@ -148,16 +149,13 @@ func (p *regexPredicate) keep(v *pq.Value) bool {
 	}
 
 	// Only alloc the string when updating the map
-	p.matches[v.String()] = matched
+	p.matches[string(b)] = matched
 
 	return matched
 }
 
 func (p *regexPredicate) KeepColumnChunk(cc pq.ColumnChunk) bool {
 	p.helper.setNewRowGroup()
-
-	// Reset match cache on each row group change
-	p.matches = make(map[string]bool, len(p.matches))
 
 	// Can we do any filtering here?
 	return true
@@ -168,6 +166,20 @@ func (p *regexPredicate) KeepValue(v pq.Value) bool {
 }
 
 func (p *regexPredicate) KeepPage(page pq.Page) bool {
+	if p.helper.newRowGroup {
+		// Reset match cache on each row group change
+		// We delay until the first page is received
+		// so we can get an accurate count of the number
+		// of distinct values for dictionary columns.
+		count := len(p.matches)
+		if d := page.Dictionary(); d != nil {
+			if d.Len() > count {
+				count = d.Len()
+			}
+		}
+		p.matches = make(map[string]bool, count)
+	}
+
 	return p.helper.keepPage(page, p.KeepValue)
 }