**What this PR does**:
Makes regex predicate faster.  Particularly the case where there are found matches because of all the map access.

Tweak originally discussed https://github.com/grafana/tempo/pull/2410#discussion_r1181532807

```
name                                                 old time/op    new time/op    delta
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-12    2.51ms ± 2%    2.45ms ± 0%   -2.52%  (p=0.000 n=9+8)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-12      72.1ms ± 2%    49.5ms ± 2%  -31.36%  (p=0.000 n=9+10)

name                                                 old speed      new speed      delta
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-12   385MB/s ± 2%   395MB/s ± 0%   +2.58%  (p=0.000 n=9+8)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-12     186MB/s ± 2%   271MB/s ± 2%  +45.69%  (p=0.000 n=9+10)

name                                                 old MB_io/op   new MB_io/op   delta
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-12      0.97 ± 0%      0.97 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-12        13.4 ± 0%      13.4 ± 0%     ~     (all equal)

name                                                 old alloc/op   new alloc/op   delta
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-12    1.10MB ± 4%    1.07MB ± 4%   -2.71%  (p=0.017 n=10+9)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-12      33.4MB ± 1%    12.5MB ± 0%  -62.73%  (p=0.000 n=10+7)

name                                                 old allocs/op  new allocs/op  delta
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-12     12.9k ± 0%     12.8k ± 0%   -0.18%  (p=0.000 n=10+7)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-12        786k ± 0%       14k ± 0%  -98.22%  (p=0.000 n=10+10)
```

**Which issue(s) this PR fixes**:
Fixes n/a

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`