@mdisibio 2023-05-17
> nice. i didn't know about that zero alloc optimization on maps

Neither did I 💡